Extract game state types/utils and harden completion pipeline
Move engine types to gameStateTypes.ts and shared stateUtils helpers to
break circular imports. Consolidate win evaluation to a single pass,
reuse playerWonGame in completion, batch achievement writes, and add
game completion idempotency checks.
