+++ packages/shared/src/engine/auction.ts
@@ -20,7 +20,8 @@ import type {
   InternalGameState,
   InternalPlayerState,
   LogEntry,
-} from "./gameStateMachine.js";
+} from "./gameStateTypes.js";
+import { deepClone, getPlayer } from "./stateUtils.js";
 import { applyWinIfThresholdCrossed } from "./winResolution.js";
 
 export type { DeclineAuctionType } from "./auctionMode.js";
@@ -37,17 +38,6 @@ export type PendingAuctionState = {
   bidDeadlineAt: number;
   settleDeadlineAt?: number;
 };
-
-function deepClone<T>(obj: T): T {
-  return JSON.parse(JSON.stringify(obj)) as T;
-}
-
-function getPlayer(
-  state: InternalGameState,
-  playerId: string,
-): InternalPlayerState | undefined {
-  return state.players.find((player) => player.playerId === playerId);
-}
 
 function auctionBidDeadline(state: InternalGameState, nowMs: number): number {
   return computeAuctionBidDeadline(nowMs, state.settings);
